fix: moved convertCamelToSnakeCase function to common and updated imports
