+++ src/common/utils.ts
@@ -289,3 +289,14 @@ export const geoContextQuery = (
     ],
   };
 };
+
+export const convertCamelToSnakeCase = (obj: Record<string, unknown>): Record<string, unknown> => {
+  const snakeCaseObj: Record<string, unknown> = {};
+  for (const key in obj) {
+    if (Object.prototype.hasOwnProperty.call(obj, key)) {
+      const snakeCaseKey = key.replace(/[A-Z]/g, (match) => `_${match.toLowerCase()}`);
+      snakeCaseObj[snakeCaseKey] = obj[key];
+    }
+  }
+  return snakeCaseObj;
+};
